SCHEDULAR/requirements.txt
pyaes
mysql-connector-python
confluent-kafka
kafka-python
schedule
prometheus-client
